add test reading/parsing funcs
